gplsone(plstools)	R Documentation

GPLS model based on NIPALS algorithm

Description

Generalized Partial least squares regresssion based on NIPALS algorithm (IWPLS).

Usage

gplsone(formula, data,subset,weights=NULL,na.action,offset,
  family=gaussian,nf = 2,method="gplsone.fit",scale=FALSE,
  tol=1e-9,deps=1e-20,nitermax=100, ...)
gplsone.fit(y,X,nf=2,family,weights=NULL,tol=1e-9,deps=1e-20,
  nitermax=100,scale=FALSE,...)
cgplsone.fit(y,X,nf=2,family,weights=NULL,tol=1e-9,deps=1e-20,
  nitermax=100,scale=FALSE,...)
## S3 method for class 'gplsone':
residuals(object,...)

Arguments

object 	 an object of class inheriting from 'gplsone'.

formula 	 an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under Details.

data 	 an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula)

family 	 a description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function. (See "family" for details of family functions.)

weights 	 an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.

subset 	 an optional vector specifying a subset of observations to be used in the fitting process.

na.action 	 a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The 'factory-fresh' default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful. Value na.pass is used to take in account the missing values.

offset 	 this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length either one or equal to the number of cases. One or more offset terms can be included in the formula instead or as well, and if both are specified their sum is used (See model.offset).

y 	 a numeric vector corresponding to the response

X 	 a numeric data.frame corresponding to explanatory variables (see model.matrix)

nf 	 an integer indicating the number of kept components

scale 	 a logical value. If scale is TRUE, X is scaled by dividing each variable by its sample standard deviation (by deafult, scale=TRUE).

method 	 the method to be used in fitting the model. The default method "plsone.fit" uses partial least squares (PLS). The only current alternative is "model.frame" which returns the model frame and does no fitting.

tol 	 positive convergence tolerance (by default, tol=1e-9).

deps 	 positive tolerance (used in the computation of convergence value,by default deps=1e-20).

nitermax 	 integer giving the maximal number of IWPLS iterations (by default, nitermax=100).

... 	 further arguments passed to or from other methods.

Details

The function 'plsone' is regression methods based on partail least square.
The default method 'plsone.fit' uses NIPALS algorithm.
The Coordinates ch,th, ph and ch correspond to Slopes equation in linear least squares regression without intercept term and can be compute with missing data.
+ additional details on the computation of the terms.
The function 'gplsone' is base on two functions: scalew (from the library ade4) and delete.intercept (from the library pls). For more details on these both methods, you can directly consult their documentations.
The functions 'cgplsone.fit' is based on C code and the function 'gpls' is programmed in R.

Value

The function 'gplsone' returns an object of class inheriting from 'plsone'.

glm 	 an object 'glm' based on the PLS components (th)

bh 	 an numeric vector which contains coefficients associated with the explanatory variables

y 	 a numeric vector corresponding to the response

x 	 a numeric data.frame corresponding to explanatory variables

ch 	 a numeric vector containing regression coefficients of y on the components 'th'

th 	 an object 'data.frame' containing the conponents 'th'.

uh 	 an object 'data.frame' containing normalised residuals for h > 1.

ph 	 an object 'data.frame' containing regression coefficient of the h_{th} columns of X_{h-1} on t(h).

wh 	 an object 'data.frame' containing regression coefficient of the h_{th} columns of X_{h-1} on y(h-1).

whx 	 an object 'data.frame' containing regression coefficent of th on explanatory variables.

nf 	 an integer indicating the number of kept components.

formula 	 the initial formula supplied.

method 	 the name of the fitter function used, currently always "plsone.fit".

call 	 generally match.call()

Note

The function 'pls' use the function 'delete.intercept' from the package pls.

References

Marx 1996

See Also

gpls, delete.intercept, glm

Examples

require(pls)
data(yarn)

[Package plstools version 1.0-6 Index]