plstools-package(plstools)	R Documentation

PLS regression based on NIPALS algoritm

Description

This package contains PLS regression functions based on NIPALS algoritm and some tools to evaluate the goodness of fit and the model quality.

Details

Package: 	 plstools

Type: 	 Package

Version: 	 1.0

Date: 	 2008-09-18

License: 	 GPL version 2 or newer

LazyLoad: 	 yes

Depends: 	 R (>= 2.7.1),ade4

Suggests: 	 ade4, boot

The function 'pls' use the function 'delete.intercept' from the package pls.

Author(s)

Pierre Bady <pierre.bady@free.fr>
Maintainer: Pierre Bady <pierre.bady@free.fr>

References

Wold H. (1966) Estimation of principal components and related models by iterative least squares. In P. Krishnaiah, editors.Multivariate Analysis, Academic Press, 391–420.
Wold H. (1975) Modeling in Complex Situations with Soft Information, Third World Congress of Econometric Society, August 21-26, Toronto, Canada.
Wold S., Martens H. & Wold H. (1983) « The multivariate calibration problem in chemistry solved by the PLS method», in Proc. Conf. Matrix Pencils, Ruhe, A. & Kågstrøm, B. (Eds), March 1982, Lecture Notes in Mathematics, Springer Verlag, Heidelberg, pp. 286- 293.
Tenenhaus M.(1998) La Regression PLS. Theorie et pratique. Technip, Paris.

See Also

pls, gpls, ade4, boot

Examples

require(pls)
data(yarn)
yarn.pls <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
plstest1 <- plsone(density ~ NIR, nf=6, data = yarn,scale=FALSE)
cor(plstest1$th,yarn.pls$scores)
cor(fitted(plstest1),yarn.pls$fitted.values[,,6])
plot(fitted(plstest1),yarn.pls$fitted.values[,,6],pch=20,panel.first=c(grid()),cex=2)
abline(0,1,col="red")
abline(0,1,col="red",lwd=2)
plstest2 <- update(plstest1,nf=2)
plstest2
plstest1
summary(plstest1)
summary(plstest2)
plot(plstest1,which.par=1:2)
plot(plstest1,which.par=c(1:2,6))

[Package plstools version 1.0-5 Index]