predict.gplsone(plstools)	R Documentation

Predict method for Generalized Partial Least Squares Fits

Description

The function 'predict.gplsone'retruns predicted values, confidence and tolerance intervals.

Usage

## S3 method for class 'gplsone':
predict(object,newdata=NULL,type="link",se.fit=FALSE,interval="none",level=0.95,...)

Arguments

object 	 Object of class inheriting from 'plsone'

newdata 	 An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

type 	 the type of prediction required. The default is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable. Thus for a default binomial model the default predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities. The "terms" option returns a matrix giving the fitted values of each term in the model formula on the linear predictor scale. The value of this argument can be abbreviated.

se.fit 	 A switch indicating if standard errors are required.

interval 	 none, confidence, prediction

level 	 Tolerance/confidence level

... 	 further arguments passed to or from other methods.

Details

add formula for tolerance and confidence interval

Value

The function 'predict.plsone'retruns ...

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

References

Tenenhaus M. (1998) La Regression PLS. Theorie et pratique. Technip, Paris.

See Also

gplsone, predict.glm

Examples

require(pls)
data(yarn)

[Package plstools version 1.0-5 Index]