summary.gplsone(plstools)	R Documentation

Summarizing Generalized Partial Least Squares model

Description

This function provides additional elements to evaluate the goodness of fit and the quality of the model (object of class inheriting from 'gplsone').

Usage

## S3 method for class 'gplsone':
summary(object, ...)
## S3 method for class 'summary.gplsone':
print(x, ...)

Arguments

object 	 an object of class inheriting from 'gplsone'

x 	 an object of class inheriting from 'summary.gplsone'

... 	 further arguments passed to or from other methods.

Details

decrire le calcul des elements ...

+ additional details on the computation of each elemnts
+ At the moment this function is very slow to compute the diagnostic elements.

Value

The function 'summary.gplsone' retruns an list of class 'summary.plsone' containing several elements associated with the model diagnostic:

bh 	 an numeric vector which contains coefficients associated with the explanatory variables

n 	 observation number

p 	 number of explanatory varaibles.

nf 	 an integer indicating the number of kept components (h).

References

Marx 1996

See Also

gplsone,summary.glm

Examples

require(pls)
data(yarn)

[Package plstools version 1.0-5 Index]