summary.plsone(plstools)	R Documentation

Summarizing Partial Least Squares regression

Description

This function provides additional elements to evaluate the goodness of fit and the quality of the model (object of class inheriting from 'plsone').

Usage

## S3 method for class 'plsone':
summary(object, ...)
## S3 method for class 'summary.plsone':
print(x,digits = max(3, getOption("digits") - 3),...)

Arguments

object 	 an object of class inheriting from 'plsone'.

x 	 an object of class inheriting from 'summary.plsone'.

digits 	 the number of significant digits to use when printing.

... 	 further arguments passed to or from other methods.

Details

decrire le calcul des elements ...

+ additional details on the computation of each elemnts
+ At the moment this function is very slow to compute the diagnostic elements.

FT2 = nf * (n * n - 1) * qf(0.95, nf, n - nf)/(n * (n - nf))
DcritX and DcritY =

Value

The function 'summary.plsone' retruns an list of class 'summary.plsone' containing several elements associated with the model diagnostic:

rx2 	 a matrix containing the R2 between x and th

ry2 	 a matrix containing the R2 between y and th

T2 	 a numeric vector containing Hotelling's T2 for each observation.

Ts2 	 a numeric vector containing adapted T2 for each observation (statistic used in SIMCA-P).

FT2 	 a numerical vector containing threshold to detected outlier with Hotelling's T2 (F-value).

rxt2 	 a matrix containing the correlations between x and th

ryt2 	 a matrix containing the correlations between y and th

DModX 	 a numeric vector containing distance from the i^{th} observation to the model in the space of the explanatory variables

DModXN 	 a numeric vector containing normaliszed values of DModX.

DcritX 	 a numerical vector containing threshold to detected the observations badly rebuilt by the model.

DModY 	 a numeric vector containing residuals of the regression of y on the 'nf' components.

DModYN 	 a numeric vector containing normaliszed values of DModY.

DcritY 	 a numerical vector containing threshold to detected the observations badly rebuilt by the model.

coefficients 	 a numeric vector containing regression coefficients of y on the components 'th'

r.squared 	 a numerical value corresponding to the R squared of linear model ('fraction of variance explained by the model',see 'summary.lm').

adj.r.squared 	 a numerical value corresponding to the adjusted R squared of linear model

bh 	 an numeric vector which contains coefficients associated with the explanatory variables

n 	 observation number

p 	 number of explanatory varaibles.

nf 	 an integer indicating the number of kept components (h).

References

Tenenhaus M. (1998) La Regression PLS. Theorie et pratique. Technip, Paris.

See Also

plsone,summary.lm

Examples

require(pls)
data(yarn)
plstest1 <- plsone(density ~ NIR, nf=6, data = yarn,scale=FALSE)
summary(plstest1)
summary(plstest1$lm)

[Package plstools version 1.0-5 Index]